Warning/deprecation fixes: std::random_shuffle->shuffle

+++ highwayhash/scalar.h
@@ -123,7 +123,7 @@ class Scalar {
   }
 
   HH_INLINE Scalar& operator>>=(const int count) {
-    if (count >= sizeof(T) * 8) {
+    if (count >= static_cast<int>(sizeof(T)) * 8) {
       v_ = 0;
     } else {
       v_ >>= count;
